+++ CHANGELOG.md
@@ -4,6 +4,21 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.0-pre.2] - 2025-01-25
+
+### Added
+
+* General APIs:
+  * FBInstant.GetLocale
+  * Player.GetASIDAsync
+* Payments APIs:
+  * FBInstant.Payments
+  * Payments.GetCatalogAsync
+  * Payments.PurchaseAsync
+  * Payments.GetPurchasesAsync
+  * Payments.ConsumePurchaseAsync
+  * Payments.OnReady
+
 ## [1.0.0-pre.1] - 2025-01-06
 
 ### Removed